Add files via upload
I’m happy to share this Zomato_DataAnalysis_PowerBi project I worked on. 

Check it out here: https://github.com/Saarthak-2000/Zomato_DataAnalysis

Objectives: To construct a consolidated and interactive Power BI report that will allow Zomato to quickly assess the required data.

Zomato is a restaurant aggregation and meal delivery service based in India. It is currently operating in several countries across the world. Zomato provides thorough information about numerous eateries as well as consumer reviews. Zomato's owners aim to find hidden irregularities in their company's data. The ultimate goal of this project is to examine the data in such a way that they can accurately assess their business performance.

The data (sample) is currently accessible in the form of a few Excel files, each of which contains information about multiple restaurants operating in a certain continent. The clients want to construct a consolidated and interactive Power BI report that will allow them to do the following:

1.	Derive data on the total number of restaurants worldwide, including continents, countries, and cities

2.	View data on a global scale with the capacity to drill down to a granular level

3.	Derive data on the restaurants with the highest average customer ratings

4.	Discover the restaurants with the lowest average costs

5.	Filter and view information on the restaurants based on:

•	Their geographical dimensions such as continent, country, and city.

•	The service they provide, such as online ordering or reservation services

•	The average rating slab by the colour.

6. Identify the restaurants with the most cuisines served

7. Design a multi-page report that suits Zomato's theme with easy navigation across sections.

8. Allow Zomato users to be able to access this information from both a web browser and a mobile device.

Aim of the project: The aim is to construct a consolidated and interactive PowerBI report that will allow Zomato to quickly assess the required data. #powerbi #powerbideveloper #powerbidashboard #dataanalysis #datavisualization